+++ source/extensions/filters/http/ext_authz/config.h
@@ -15,11 +15,11 @@ namespace ExtAuthz {
  */
 class ExtAuthzFilterConfig : public Server::Configuration::NamedHttpFilterConfigFactory {
 public:
-  Server::Configuration::HttpFilterFactoryCb
+  Http::FilterFactoryCb
   createFilterFactory(const Json::Object& json_config, const std::string&,
                       Server::Configuration::FactoryContext& context) override;
 
-  Server::Configuration::HttpFilterFactoryCb
+  Http::FilterFactoryCb
   createFilterFactoryFromProto(const Protobuf::Message& proto_config,
                                const std::string& stats_prefix,
                                Server::Configuration::FactoryContext& context) override;
@@ -32,7 +32,7 @@ public:
   std::string name() override { return HttpFilterNames::get().EXT_AUTHORIZATION; }
 
 private:
-  Server::Configuration::HttpFilterFactoryCb
+  Http::FilterFactoryCb
   createFilter(const envoy::config::filter::http::ext_authz::v2alpha::ExtAuthz& proto_config,
                const std::string& stats_prefix, Server::Configuration::FactoryContext& context);
 };