+++ source/extensions/filters/http/grpc_web/config.h
@@ -12,8 +12,8 @@ namespace GrpcWeb {
 
 class GrpcWebFilterConfig : public Common::EmptyHttpFilterConfig {
 public:
-  Server::Configuration::HttpFilterFactoryCb
-  createFilter(const std::string&, Server::Configuration::FactoryContext& context) override;
+  Http::FilterFactoryCb createFilter(const std::string&,
+                                     Server::Configuration::FactoryContext& context) override;
 
   std::string name() override { return HttpFilterNames::get().GRPC_WEB; }
 };