+++ test/mocks/router/mocks.h
@@ -297,15 +297,15 @@ public:
   ~MockRouteConfigProviderManager();
 
   MOCK_METHOD0(routeConfigProviders, std::vector<RouteConfigProviderSharedPtr>());
-  MOCK_METHOD5(getRdsRouteConfigProvider,
+  MOCK_METHOD3(getRdsRouteConfigProvider,
                RouteConfigProviderSharedPtr(
                    const envoy::config::filter::network::http_connection_manager::v2::Rds& rds,
-                   Upstream::ClusterManager& cm, Stats::Scope& scope,
-                   const std::string& stat_prefix, Init::Manager& init_manager));
-  MOCK_METHOD3(getStaticRouteConfigProvider,
-               RouteConfigProviderSharedPtr(const envoy::api::v2::RouteConfiguration& route_config,
-                                            Runtime::Loader& runtime,
-                                            Upstream::ClusterManager& cm));
+                   Server::Configuration::FactoryContext& factory_context,
+                   const std::string& stat_prefix));
+  MOCK_METHOD2(
+      getStaticRouteConfigProvider,
+      RouteConfigProviderSharedPtr(const envoy::api::v2::RouteConfiguration& route_config,
+                                   Server::Configuration::FactoryContext& factory_context));
 
   MOCK_METHOD1(removeRouteConfigProvider, void(const std::string& identifier));
 };